Add files via upload
The NCDC (Nigeria Centre for Disease Control) dataset is based on COVID-19 data from a number of states across Nigeria, it depicts the number of confirmed Covid Cases, deaths and recoveries across the states. The data is further broken down in these categories according to gender, age and states.
With total confirmed cases at 266,000; 259,000 recoveries, 4063 active cases, and 3155 deaths. The total recoveries distributed across the states, depict the highest number of recoveries occurring in Lagos state, FCT and Port Harcourt. The record of 3155 deaths are split across gender in the proportion of 93.5%  male and 6.5% female.